refactor(pipeline/scrape): extract TxnMapper module
Phase 5 — Commit 6 of 9 (pipeline-decoupling-master-2026-05-28).

Carves the single-record txn-mapping concern out of
ScrapeAutoMapper.ts into a dedicated sub-module under the
per-cluster max-lines:150 eff cap (96 eff).

TxnMapper/TxnMapper.ts (96 eff) owns:

- coerceIdentifier — sanitises raw identifier hits, rejects
  the `0` / empty-string sentinels Beinleumi emits on pending
  out-of-statement rows so the dedup hash never collides.
- SHEKEL_ALIASES + normalizeCurrency — converts every
  WK-listed shekel spelling to canonical `ILS`.
- isVoidedTransaction — guards extractTransactions /
  extractTransactionsForCard against dealSumType=1 voided rows
  and voucherNumberRatz=000000000 invalid stubs.
- maybeNegateAmount — applies the card-vs-bank sign
  convention (Isracard/Amex report positive amounts for charges).
- resolveAmount — single-field or split debit/credit netting
  fallback so banks that emit (debit, credit) instead of one
  signed `amount` still map correctly.
- applyDirectionWk — flips sign when WK.direction reports a
  debit indicator.
- isMappableTxn — rejects records with empty date or NaN
  amount BEFORE they reach the autoMapTransaction output, so
  bad rows are logged and dropped at the auto-mapper boundary
  instead of silently disappearing later in deduplicateTxns.
- autoMapTransaction — composes the helpers into a single
  `raw → ITransaction | false` mapping.

Ordering note: TxnMapper extracted BEFORE ContainerPicker
(Commit 7) because ContainerPicker's extractTransactions and
extractTransactionsForCard helpers call autoMapTransaction +
isVoidedTransaction. Pulling the mapper out first lets
ContainerPicker import them via a clean `../TxnMapper/TxnMapper.js`
path instead of reaching back into the in-flux ScrapeAutoMapper
parent.

ScrapeAutoMapper.ts goes from 295 → 214 eff LoC (-81). The
public re-export block at the bottom continues to surface
`autoMapTransaction` so no consumer needs to change its
import path.

No production behaviour change. All 66 ScrapeAutoMapper-related
tests still pass unchanged. Husky 15-gate hook green.
